+++ s/demo/elements/slate-silver.ts
@@ -1,11 +1,11 @@
 
 import {html} from "lit"
 import {NestingOuter} from "../views/nesting.js"
-import {SilverElement} from "../../element/silver.js"
+import {LightElement} from "../../element/silver.js"
 import {QuartzTripler} from "../views/quartz-tripler.js"
 import {ObsidianQuadrupler} from "../views/obsidian-quadrupler.js"
 
-export class SlateSilver extends SilverElement {
+export class SlateSilver extends LightElement {
 	render() {
 		return html`
 			${QuartzTripler(1)}