+++ s/nexus/parts/use/parts/use_gold.ts
@@ -1,15 +1,15 @@
 
 import {UseShadow} from "./use_shadow.js"
 import {Context} from "../../../context.js"
-import {GoldElement} from "../../../../element/gold.js"
+import {ShadowElement} from "../../../../element/gold.js"
 import {SetupAttrs, setup_use_attrs} from "../../setup_use_attrs.js"
 
-export class UseGold<C extends Context = Context> extends UseShadow<C, GoldElement> {
+export class UseGold<C extends Context = Context> extends UseShadow<C, ShadowElement> {
 
 	readonly attrs: SetupAttrs
 
 	constructor(
-			element: GoldElement,
+			element: ShadowElement,
 			shadow: ShadowRoot,
 			rerender: () => void,
 			context: C,