+++ s/nexus/parts/use/parts/use_silver.ts
@@ -1,18 +1,18 @@
 
 import {Use} from "./use.js"
 import {Context} from "../../../context.js"
-import {SilverElement} from "../../../../element/silver.js"
+import {LightElement} from "../../../../element/silver.js"
 import {SetupAttrs, setup_use_attrs} from "../../setup_use_attrs.js"
 
 export class UseSilver<C extends Context = Context> extends Use<C> {
 
-	#element: SilverElement
+	#element: LightElement
 	get element() { return this.#element }
 
 	readonly attrs: SetupAttrs
 
 	constructor(
-			element: SilverElement,
+			element: LightElement,
 			rerender: () => void,
 			context: C,
 		) {